Updated LazyPreparedStatement to use a dynamic proxy.
git-svn-id: https://fluid.surelogic.com/svn/eng/trunk/sierra-jdbc@2095 d02ae147-ee12-4a35-9452-91c89e8b2592

+++ src/com/surelogic/sierra/jdbc/project/ClientProjectManager.java
@@ -139,7 +139,11 @@ public class ClientProjectManager extends ProjectManager {
 				monitor.subTask("Deleting scans for project " + projectName);
 			scanManager.deleteScans(getProjectScans(rec.getId()), monitor);
 			findingManager.deleteFindings(projectName, monitor);
-			rec.delete();
+			if(monitor != null){
+				if(!monitor.isCanceled()) {
+					rec.delete();					
+				}
+			}
 		}
 	}
